requirements.txt
python-telegram-bot>=20.1
python-dotenv>=0.21.1